refactor(cloud-init): split orchestrate/load/store responsibilities
