+++ src/wsl/cloud_init/store.rs
@@ -1,39 +1,24 @@
-use crate::wsl::helpers::userprofile::resolve_userprofile_dir;
 use log::{info, warn};
-use std::path::PathBuf;
+use std::path::Path;
 
-pub fn store(hostname: &str, rendered: &str, dry_run: bool, debug: bool) -> anyhow::Result<PathBuf> {
-    let target_file = create_cloud_init_target(hostname, dry_run)?;
-    if !dry_run {
-        std::fs::write(&target_file, rendered)?;
-    }
-    if debug {
-        debug_cloud_init(rendered, hostname);
-    }
-    Ok(target_file)
+pub fn store(target_file: &Path, rendered: &str) -> anyhow::Result<()> {
+    let target_dir = target_file
+        .parent()
+        .ok_or_else(|| anyhow::anyhow!("cloud-init target missing parent directory"))?;
+    std::fs::create_dir_all(target_dir)?;
+    std::fs::write(target_file, rendered)?;
+    Ok(())
 }
 
-// Determine the target path for the cloud-init user-data file based on the hostname.
-fn create_cloud_init_target(hostname: &str, dry_run: bool) -> anyhow::Result<PathBuf> {
-    if dry_run {
-        info!("🧪 Dry run: cloud-init target would be created at: {}", hostname);
-        return Ok(PathBuf::from(format!("{}.user-data", hostname)));
-    }
-    let userprofile = resolve_userprofile_dir()?;
-    let target_dir = userprofile.join(".cloud-init");
-    std::fs::create_dir_all(&target_dir)?;
-    Ok(target_dir.join(format!("{}.user-data", hostname)))
-}
-
-fn debug_cloud_init(rendered: &str, hostname: &str) {
-    let debug_path = match std::env::current_dir() {
-        Ok(dir) => dir.join(format!("cloud-init.{}.user-data", hostname)),
+pub fn copy_debug_to_current_dir(hostname: &str, rendered: &str) {
+    let debug_dir = match std::env::current_dir() {
+        Ok(dir) => dir,
         Err(err) => {
             warn!("☁️ Cloud-init debug copy skipped (cwd error): {err}");
             return;
         }
     };
-
+    let debug_path = debug_dir.join(format!("cloud-init.{}.user-data", hostname));
     if let Err(err) = std::fs::write(&debug_path, rendered) {
         warn!("☁️ Cloud-init debug copy skipped (write error): {}", err);
     } else {
